split mathics.doc.common_doc
