+++ mathics/doc/common_doc.py
@@ -22,113 +22,39 @@ As with reading in data, final assembly to a LaTeX file or running
 documentation tests is done elsewhere.
 
 FIXME: This code should be replaced by Sphinx and autodoc.
-Things are such a mess, that it is too difficult to contemplate this right now. Also there
-higher-priority flaws that are more more pressing.
-In the shorter, we might we move code for extracting printing to a separate package.
+Things are such a mess, that it is too difficult to contemplate this right now. 
+Also there higher-priority flaws that are more more pressing.
+In the shorter, we might we move code for extracting printing to a
+separate package.
 """
 
-import importlib
 import logging
 import os.path as osp
 import pkgutil
 import re
-from os import environ, getenv, listdir
+from os import environ, listdir
 from types import ModuleType
-from typing import Callable, Iterator, List, Optional, Tuple
+from typing import Iterator, List, Optional, Tuple
 
 from mathics import settings
 from mathics.core.builtin import check_requires_list
-from mathics.core.evaluation import Message, Print
 from mathics.core.load_builtin import (
     builtins_by_module as global_builtins_by_module,
     mathics3_builtins_modules,
 )
 from mathics.core.util import IS_PYPY
+from mathics.doc.doctests import (
+    DocumentationEntry,
+    Tests,
+    filter_comments,
+    parse_docstring_to_DocumentationEntry_items,
+)
 from mathics.doc.utils import slugify
 from mathics.eval.pymathics import pymathics_builtins_by_module, pymathics_modules
 
-# These are all the XML/HTML-like tags that documentation supports.
-ALLOWED_TAGS = (
-    "dl",
-    "dd",
-    "dt",
-    "em",
-    "url",
-    "ul",
-    "i",
-    "ol",
-    "li",
-    "con",
-    "console",
-    "img",
-    "imgpng",
-    "ref",
-    "subsection",
-)
-ALLOWED_TAGS_RE = dict(
-    (allowed, re.compile("&lt;(%s.*?)&gt;" % allowed)) for allowed in ALLOWED_TAGS
-)
-
-# This string is used, so we can indicate a trailing blank at the end of a line by
-# adding this string to the end of the line which gets stripped off.
-# Some editors and formatters like to strip off trailing blanks at the ends of lines.
-END_LINE_SENTINAL = "#<--#"
-
-# The regular expressions below (strings ending with _RE
-# pull out information from docstring or text in a file. Ghetto parsing.
-
 CHAPTER_RE = re.compile('(?s)<chapter title="(.*?)">(.*?)</chapter>')
-CONSOLE_RE = re.compile(r"(?s)<(?P<tag>con|console)>(?P<content>.*?)</(?P=tag)>")
-DL_ITEM_RE = re.compile(
-    r"(?s)<(?P<tag>d[td])>(?P<content>.*?)(?:</(?P=tag)>|)\s*(?:(?=<d[td]>)|$)"
-)
-DL_RE = re.compile(r"(?s)<dl>(.*?)</dl>")
-HYPERTEXT_RE = re.compile(
-    r"(?s)<(?P<tag>em|url)>(\s*:(?P<text>.*?):\s*)?(?P<content>.*?)</(?P=tag)>"
-)
-IMG_PNG_RE = re.compile(
-    r'<imgpng src="(?P<src>.*?)" title="(?P<title>.*?)" label="(?P<label>.*?)">'
-)
-IMG_RE = re.compile(
-    r'<img src="(?P<src>.*?)" title="(?P<title>.*?)" label="(?P<label>.*?)">'
-)
-# Preserve space before and after in-line code variables.
-LATEX_RE = re.compile(r"(\s?)\$(\w+?)\$(\s?)")
-
-LIST_ITEM_RE = re.compile(r"(?s)<li>(.*?)(?:</li>|(?=<li>)|$)")
-LIST_RE = re.compile(r"(?s)<(?P<tag>ul|ol)>(?P<content>.*?)</(?P=tag)>")
-MATHICS_RE = re.compile(r"(?<!\\)\'(.*?)(?<!\\)\'")
-
-PYTHON_RE = re.compile(r"(?s)<python>(.*?)</python>")
-QUOTATIONS_RE = re.compile(r"\"([\w\s,]*?)\"")
-REF_RE = re.compile(r'<ref label="(?P<label>.*?)">')
 SECTION_RE = re.compile('(?s)(.*?)<section title="(.*?)">(.*?)</section>')
-SPECIAL_COMMANDS = {
-    "LaTeX": (r"<em>LaTeX</em>", r"\LaTeX{}"),
-    "Mathematica": (
-        r"<em>Mathematica</em>&reg;",
-        r"\emph{Mathematica}\textregistered{}",
-    ),
-    "Mathics": (r"<em>Mathics3</em>", r"\emph{Mathics3}"),
-    "Mathics3": (r"<em>Mathics3</em>", r"\emph{Mathics3}"),
-    "Sage": (r"<em>Sage</em>", r"\emph{Sage}"),
-    "Wolfram": (r"<em>Wolfram</em>", r"\emph{Wolfram}"),
-    "skip": (r"<br /><br />", r"\bigskip"),
-}
-SUBSECTION_END_RE = re.compile("</subsection>")
 SUBSECTION_RE = re.compile('(?s)<subsection title="(.*?)">')
-
-TESTCASE_RE = re.compile(
-    r"""(?mx)^  # re.MULTILINE (multi-line match)
-                # and re.VERBOSE (readable regular expressions
-        ((?:.|\n)*?)
-        ^\s+([>#SX])>[ ](.*)  # test-code indicator
-        ((?:\n\s*(?:[:|=.][ ]|\.).*)*)  # test-code results"""
-)
-TESTCASE_OUT_RE = re.compile(r"^\s*([:|=])(.*)$")
-
-# Used for getting test results by test expresson and chapter/section information.
-test_result_map = {}
 
 # Debug flags.
 
@@ -159,62 +85,13 @@ def get_module_doc(module: ModuleType) -> Tuple[str, str]:
         title = doc.splitlines()[0]
         text = "\n".join(doc.splitlines()[1:])
     else:
-        # FIXME: Extend me for Mathics3 modules.
         title = module.__name__
-        for prefix in ("mathics.builtin.", "mathics.optional."):
+        for prefix in ("mathics.builtin.", "mathics.optional.", "pymathics."):
             if title.startswith(prefix):
                 title = title[len(prefix) :]
         title = title.capitalize()
         text = ""
     return title, text
-
-
-def get_results_by_test(test_expr: str, full_test_key: list, doc_data: dict) -> dict:
-    """
-    Sometimes test numbering is off, either due to bugs or changes since the
-    data was read.
-
-    Here, we compensate for this by looking up the test by its chapter and section name
-    portion stored in `full_test_key` along with the and the test expression data
-    stored in `test_expr`.
-
-    This new key is looked up in `test_result_map` its value is returned.
-
-    `doc_data` is only first time this is called to populate `test_result_map`.
-    """
-
-    # Strip off the test index form new key with this and the test string.
-    # Add to any existing value for that "result". This is now what we want to
-    # use as a tee in test_result_map to look for.
-    test_section = list(full_test_key)[:-1]
-    search_key = tuple(test_section)
-
-    if not test_result_map:
-        # Populate test_result_map from doc_data
-        for key, result in doc_data.items():
-            test_section = list(key)[:-1]
-            new_test_key = tuple(test_section)
-            next_result = test_result_map.get(new_test_key, None)
-            if next_result is None:
-                next_result = [result]
-            else:
-                next_result.append(result)
-
-            test_result_map[new_test_key] = next_result
-
-    results = test_result_map.get(search_key, None)
-    result = {}
-    if results:
-        for result_candidate in results:
-            if result_candidate["query"] == test_expr:
-                if result:
-                    # Already found something
-                    print(f"Warning, multiple results appear under {search_key}.")
-                    return {}
-                else:
-                    result = result_candidate
-
-    return result
 
 
 def get_submodule_names(obj) -> list:
@@ -252,14 +129,6 @@ def get_submodule_names(obj) -> list:
     return modpkgs
 
 
-def filter_comments(doc: str) -> str:
-    """Remove docstring documentation comments. These are lines
-    that start with ##"""
-    return "\n".join(
-        line for line in doc.splitlines() if not line.lstrip().startswith("##")
-    )
-
-
 def get_doc_name_from_module(module) -> str:
     """
     Get the title associated to the module.
@@ -277,35 +146,13 @@ def get_doc_name_from_module(module) -> str:
     return name
 
 
-POST_SUBSTITUTION_TAG = "_POST_SUBSTITUTION%d_"
-
-
-def pre_sub(regexp, text: str, repl_func):
-    post_substitutions = []
-
-    def repl_pre(match):
-        repl = repl_func(match)
-        index = len(post_substitutions)
-        post_substitutions.append(repl)
-        return POST_SUBSTITUTION_TAG % index
-
-    text = regexp.sub(repl_pre, text)
-
-    return text, post_substitutions
-
-
-def post_sub(text: str, post_substitutions) -> str:
-    for index, sub in enumerate(post_substitutions):
-        text = text.replace(POST_SUBSTITUTION_TAG % index, sub)
-    return text
-
-
 def skip_doc(cls) -> bool:
     """Returns True if we should skip cls in docstring extraction."""
     return cls.__name__.endswith("Box") or (hasattr(cls, "no_doc") and cls.no_doc)
 
 
 def skip_module_doc(module, must_be_skipped) -> bool:
+    """True if the module should not be included in the documentation"""
     return (
         module.__doc__ is None
         or module in must_be_skipped
@@ -313,209 +160,6 @@ def skip_module_doc(module, must_be_skipped) -> bool:
         or hasattr(module, "no_doc")
         and module.no_doc
     )
-
-
-def parse_docstring_to_DocumentationEntry_items(
-    doc: str,
-    test_collection_constructor: Callable,
-    test_case_constructor: Callable,
-    text_constructor: Callable,
-    key_part=None,
-) -> list:
-    """
-    This parses string `doc` (using regular expressions) into Python objects.
-    The function returns a list of ``DocText`` and ``DocTests`` objects which
-    are contained in a ``DocumentationElement``.
-
-    test_collection_constructor() is the class constructor call to create an object for the test collection.
-    Each test is created via test_case_constructor().
-    Text within the test is stored via text_constructor.
-    """
-    # This function is used to populate a ``DocumentEntry`` element, that
-    # in principle is not associated to any container (``DocChapter``/``DocSection``/``DocSubsection``)
-    # of the documentation system.
-    #
-    # The ``key_part`` parameter was used to set the ``key`` of the ``DocTest`` elements. This attribute
-    # should be set just after the  ``DocumentationEntry`` (to which the tests belongs) is associated
-    # to a container, by calling  ``container.set_parent_path``.
-    # However, the parameter is still used in MathicsDjango, so let's keep it and discard its value.
-    #
-    if key_part:
-        logging.warning("``key_part`` is deprecated. Its value is discarded.")
-
-    # Remove commented lines.
-    doc = filter_comments(doc).strip(r"\s")
-
-    # Remove leading <dl>...</dl>
-    # doc = DL_RE.sub("", doc)
-
-    # pre-substitute Python code because it might contain tests
-    doc, post_substitutions = pre_sub(
-        PYTHON_RE, doc, lambda m: "<python>%s</python>" % m.group(1)
-    )
-
-    # HACK: Artificially construct a last testcase to get the "intertext"
-    # after the last (real) testcase. Ignore the test, of course.
-    doc += "\n >> test\n = test"
-    testcases = TESTCASE_RE.findall(doc)
-
-    tests = None
-    items = []
-    for index in range(len(testcases)):
-        testcase = list(testcases[index])
-        text = testcase.pop(0).strip()
-        if text:
-            if tests is not None:
-                items.append(tests)
-                tests = None
-            text = post_sub(text, post_substitutions)
-            items.append(text_constructor(text))
-            tests = None
-        if index < len(testcases) - 1:
-            test = test_case_constructor(index, testcase, None)
-            if tests is None:
-                tests = test_collection_constructor()
-            tests.tests.append(test)
-
-    # If the last block in the loop was not a Text block, append the
-    # last set of tests.
-    if tests is not None:
-        items.append(tests)
-        tests = None
-    return items
-
-
-class DocTest:
-    """
-    Class to hold a single doctest.
-
-    DocTest formatting rules:
-
-    * `>>` Marks test case; it will also appear as part of
-           the documentation.
-    * `#>` Marks test private or one that does not appear as part of
-           the documentation.
-    * `X>` Shows the example in the docs, but disables testing the example.
-    * `S>` Shows the example in the docs, but disables testing if environment
-           variable SANDBOX is set.
-    * `=`  Compares the result text.
-    * `:`  Compares an (error) message.
-      `|`  Prints output.
-    """
-
-    def __init__(
-        self, index: int, testcase: List[str], key_prefix: Optional[tuple] = None
-    ):
-        def strip_sentinal(line: str):
-            """Remove END_LINE_SENTINAL from the end of a line if it appears.
-
-            Some editors like to strip blanks at the end of a line.
-            Since the line ends in END_LINE_SENTINAL which isn't blank,
-            any blanks that appear before will be preserved.
-
-            Some tests require some lines to be blank or entry because
-            Mathics3 output can be that way
-            """
-            if line.endswith(END_LINE_SENTINAL):
-                line = line[: -len(END_LINE_SENTINAL)]
-
-            # Also remove any remaining trailing blanks since that
-            # seems *also* what we want to do.
-            return line.strip()
-
-        self.index = index
-        self.outs = []
-        self.result = None
-
-        # Private test cases are executed, but NOT shown as part of the docs
-        self.private = testcase[0] == "#"
-
-        # Ignored test cases are NOT executed, but shown as part of the docs
-        # Sandboxed test cases are NOT executed if environment SANDBOX is set
-        if testcase[0] == "X" or (testcase[0] == "S" and getenv("SANDBOX", False)):
-            self.ignore = True
-            # substitute '>' again so we get the correct formatting
-            testcase[0] = ">"
-        else:
-            self.ignore = False
-
-        self.test = strip_sentinal(testcase[1])
-        self._key = key_prefix + (index,) if key_prefix else None
-
-        outs = testcase[2].splitlines()
-        for line in outs:
-            line = strip_sentinal(line)
-            if line:
-                if line.startswith("."):
-                    text = line[1:]
-                    if text.startswith(" "):
-                        text = text[1:]
-                    text = "\n" + text
-                    if self.result is not None:
-                        self.result += text
-                    elif self.outs:
-                        self.outs[-1].text += text
-                    continue
-
-                match = TESTCASE_OUT_RE.match(line)
-                if not match:
-                    continue
-                symbol, text = match.group(1), match.group(2)
-                text = text.strip()
-                if symbol == "=":
-                    self.result = text
-                elif symbol == ":":
-                    out = Message("", "", text)
-                    self.outs.append(out)
-                elif symbol == "|":
-                    out = Print(text)
-                    self.outs.append(out)
-
-    def __str__(self) -> str:
-        return self.test
-
-    @property
-    def key(self):
-        return self._key if hasattr(self, "_key") else None
-
-    @key.setter
-    def key(self, value):
-        assert self.key is None
-        self._key = value
-        return self._key
-
-
-# Tests has to appear before Documentation which uses it.
-# FIXME: Turn into a NamedTuple? Or combine with another class?
-class Tests:
-    """
-    A group of tests in the same section or subsection.
-    """
-
-    def __init__(
-        self,
-        part_name: str,
-        chapter_name: str,
-        section_name: str,
-        doctests: List[DocTest],
-        subsection_name: Optional[str] = None,
-    ):
-        self.part = part_name
-        self.chapter = chapter_name
-        self.section = section_name
-        self.subsection = subsection_name
-        self.tests = doctests
-        self._key = None
-
-    @property
-    def key(self):
-        return self._key
-
-    @key.setter
-    def key(self, value):
-        assert self._key is None
-        self._key = value
-        return self._key
 
 
 # DocSection has to appear before DocGuideSection which uses it.
@@ -653,41 +297,24 @@ class DocGuideSection(DocSection):
         submodule,
         installed: bool = True,
     ):
-        # Notice that we need the documentation object, to have access
-        # to the suitable subclass of DocumentationElement.
-        documentation = chapter.part.doc
-        self.chapter = chapter
-        self.doc = documentation.doc_class(text, title, None)
-        self.in_guide = False
-        self.installed = installed
+        super().__init__(chapter, title, text, None, installed, False)
         self.section = submodule
-        self.slug = slugify(title)
-        self.subsections = []
-        self.subsections_by_slug = {}
-        self.title = title
 
-        self.doc.set_parent_path(self)
-
-        # FIXME: Sections never are operators. Subsections can have
-        # operators though.  Fix up the view and searching code not to
-        # look for the operator field of a section.
-        self.operator = False
-
-        if text.count("<dl>") != text.count("</dl>"):
-            raise ValueError(
-                "Missing opening or closing <dl> tag in "
-                "{} documentation".format(title)
-            )
         if MATHICS_DEBUG_DOC_BUILD:
             print("    DEBUG Creating Guide Section", title)
-        chapter.sections_by_slug[self.slug] = self
 
     # FIXME: turn into a @property tests?
     def get_tests(self):
+        """
+        Tests included in a Guide.
+        """
         # FIXME: The below is a little weird for Guide Sections.
         # Figure out how to make this clearer.
         # A guide section's subsection are Sections without the Guide.
         # it is *their* subsections where we generally find tests.
+        #
+        # Currently, this is not called in docpipeline or in making
+        # the LaTeX documentation.
         for section in self.subsections:
             if not section.installed:
                 continue
@@ -747,31 +374,6 @@ class DocPart:
         )
 
 
-class DocTests:
-    """
-    A bunch of consecutive ``DocTest`` objects extracted from a Builtin docstring.
-    """
-
-    def __init__(self):
-        self.tests = []
-        self.text = ""
-
-    def get_tests(self) -> list:
-        """
-        Returns lists test objects.
-        """
-        return self.tests
-
-    def is_private(self) -> bool:
-        return all(test.private for test in self.tests)
-
-    def __str__(self) -> str:
-        return "\n".join(str(test) for test in self.tests)
-
-    def test_indices(self) -> List[int]:
-        return [test.index for test in self.tests]
-
-
 class Documentation:
     """
     `Documentation` describes an object containing the whole documentation system.
@@ -799,15 +401,25 @@ class Documentation:
     the elements of the subsequent terms in the hierarchy.
     """
 
-    def __init__(self):
+    def __init__(self, title: str = "Title", doc_dir: str = ""):
+        """
+        Parameters
+        ----------
+        title : str, optional
+            The title of the Documentation. The default is "Title".
+        doc_dir : str, optional
+            The path where the sources can be loaded. The default is "",
+            meaning that no sources must be loaded.
+        """
         # This is a way to load the default classes
         # without defining these attributes as class
         # attributes.
         self._set_classes()
-        self.parts = []
         self.appendix = []
+        self.doc_dir = doc_dir
+        self.parts = []
         self.parts_by_slug = {}
-        self.title = "Title"
+        self.title = title
 
     def _set_classes(self):
         """
@@ -1012,7 +624,7 @@ class Documentation:
                 guide_section.subsections.append(section)
 
                 builtins = builtins_by_module.get(submodule.__name__, [])
-                subsections = [builtin for builtin in builtins]
+                subsections = list(builtins)
                 for instance in subsections:
                     if hasattr(instance, "no_doc") and instance.no_doc:
                         continue
@@ -1042,6 +654,10 @@ class Documentation:
         return chapter
 
     def doc_sections(self, sections, modules_seen, chapter):
+        """
+        Load sections from a list of mathics builtins.
+        """
+
         for instance in sections:
             if instance not in modules_seen and (
                 not hasattr(instance, "no_doc") or not instance.no_doc
@@ -1062,15 +678,18 @@ class Documentation:
                 modules_seen.add(instance)
 
     def get_part(self, part_slug):
+        """return a section from part key"""
         return self.parts_by_slug.get(part_slug)
 
     def get_chapter(self, part_slug, chapter_slug):
+        """return a section from part and chapter keys"""
         part = self.parts_by_slug.get(part_slug)
         if part:
             return part.chapters_by_slug.get(chapter_slug)
         return None
 
     def get_section(self, part_slug, chapter_slug, section_slug):
+        """return a section from part, chapter and section keys"""
         part = self.parts_by_slug.get(part_slug)
         if part:
             chapter = part.chapters_by_slug.get(chapter_slug)
@@ -1079,6 +698,10 @@ class Documentation:
         return None
 
     def get_subsection(self, part_slug, chapter_slug, section_slug, subsection_slug):
+        """
+        return a section from part, chapter, section and subsection
+        keys
+        """
         part = self.parts_by_slug.get(part_slug)
         if part:
             chapter = part.chapters_by_slug.get(chapter_slug)
@@ -1144,14 +767,11 @@ class Documentation:
                             tests = section.doc.get_tests()
                             if tests:
                                 yield Tests(
-                                    part.title, chapter.title, section.title, tests
+                                    part.title,
+                                    chapter.title,
+                                    section.title,
+                                    tests,
                                 )
-                                pass
-                            pass
-                        pass
-                    pass
-                pass
-            pass
         return
 
     def load_documentation_sources(self):
@@ -1222,7 +842,11 @@ class Documentation:
         return
 
     def load_part_from_file(
-        self, filename: str, title: str, chapter_order: int, is_appendix: bool = False
+        self,
+        filename: str,
+        title: str,
+        chapter_order: int,
+        is_appendix: bool = False,
     ) -> int:
         """Load a markdown file as a part of the documentation"""
         part = self.part_class(self, title)
@@ -1249,8 +873,6 @@ class Documentation:
                             text,
                         )
                         section.subsections.append(subsection)
-                        pass
-                    pass
                 else:
                     section = None
                 if not chapter.doc:
@@ -1394,158 +1016,22 @@ class MathicsMainDocumentation(Documentation):
     """
 
     def __init__(self):
-        super().__init__()
-
-        self.doc_dir = settings.DOC_DIR
+        super().__init__(title="Mathics Main Documentation", doc_dir=settings.DOC_DIR)
         self.doctest_latex_pcl_path = settings.DOCTEST_LATEX_DATA_PCL
         self.pymathics_doc_loaded = False
         self.doc_data_file = settings.get_doctest_latex_data_path(
             should_be_readable=True
         )
-        self.title = "Mathics Main Documentation"
 
     def gather_doctest_data(self):
         """
         Populates the documentatation.
         (deprecated)
         """
-        logging.warn(
+        logging.warning(
             "gather_doctest_data is deprecated. Use load_documentation_sources"
         )
         return self.load_documentation_sources()
-
-
-class DocText:
-    """
-    Class to hold some (non-test) text.
-
-    Some of the kinds of tags you may find here are showin in global ALLOWED_TAGS.
-    Some text may be marked with surrounding "$" or "'".
-
-    The code here however does not make use of any of the tagging.
-
-    """
-
-    def __init__(self, text):
-        self.text = text
-
-    def __str__(self) -> str:
-        return self.text
-
-    def get_tests(self) -> list:
-        """
-        Return tests in a DocText item - there never are any.
-        """
-        return []
-
-    def is_private(self) -> bool:
-        return False
-
-    def test_indices(self) -> List[int]:
-        return []
-
-
-# Former XMLDoc
-class DocumentationEntry:
-    """
-    A class to hold the content of a documentation entry,
-    in our internal markdown-like format data.
-
-    Describes the contain of an entry in the documentation system, as a
-    sequence (list) of items of the clase  `DocText` and `DocTests`.
-    ``DocText`` items contains an internal XML-like formatted text. ``DocTests`` entries
-    contain one or more `DocTest` element.
-    Each level of the Documentation hierarchy contains an XMLDoc, describing the
-    content after the title and before the elements of the next level. For example,
-    in ``DocChapter``, ``DocChapter.doc`` contains the text coming after the title
-    of the chapter, and before the sections in `DocChapter.sections`.
-    Specialized classes like LaTeXDoc or and DjangoDoc provide methods for
-    getting formatted output. For LaTeXDoc ``latex()`` is added while for
-    DjangoDoc ``html()`` is added
-    Mathics core also uses this in getting usage strings (`??`).
-
-    """
-
-    def __init__(self, doc_str: str, title: str, section: Optional[DocSection] = None):
-        self._set_classes()
-        self.title = title
-        self.path = None
-        if section:
-            chapter = section.chapter
-            part = chapter.part
-            # Note: we elide section.title
-            key_prefix = (part.title, chapter.title, title)
-        else:
-            key_prefix = None
-
-        self.rawdoc = doc_str
-        self.items = parse_docstring_to_DocumentationEntry_items(
-            self.rawdoc,
-            self.docTest_collection_class,
-            self.docTest_class,
-            self.docText_class,
-            None,
-        )
-
-    def _set_classes(self):
-        """
-        Tells to the initializator the classes to be used to build the items.
-        This must be overloaded by the daughter classes.
-        """
-        if not hasattr(self, "docTest_collection_class"):
-            self.docTest_collection_class = DocTests
-            self.docTest_class = DocTest
-            self.docText_class = DocText
-
-    def __str__(self) -> str:
-        return "\n\n".join(str(item) for item in self.items)
-
-    def text(self) -> str:
-        # used for introspection
-        # TODO parse XML and pretty print
-        # HACK
-        item = str(self.items[0])
-        item = "\n".join(line.strip() for line in item.split("\n"))
-        item = item.replace("<dl>", "")
-        item = item.replace("</dl>", "")
-        item = item.replace("<dt>", "  ")
-        item = item.replace("</dt>", "")
-        item = item.replace("<dd>", "    ")
-        item = item.replace("</dd>", "")
-        item = "\n".join(line for line in item.split("\n") if not line.isspace())
-        return item
-
-    def get_tests(self) -> list:
-        tests = []
-        for item in self.items:
-            tests.extend(item.get_tests())
-        return tests
-
-    def set_parent_path(self, parent):
-        """Set the parent path"""
-        self.path = None
-        path = []
-        while hasattr(parent, "parent"):
-            path = [parent.title] + path
-            parent = parent.parent
-
-        if hasattr(parent, "title"):
-            path = [parent.title] + path
-
-        if path:
-            self.path = path
-            # Set the key on each test
-            for test in self.get_tests():
-                assert test.key is None
-                # For backward compatibility, we need
-                # to reduce this to three fields.
-                # TODO: remove me and ensure that this
-                # works here and in Mathics Django
-                if len(path) > 3:
-                    path = path[:2] + [path[-1]]
-                test.key = tuple(path) + (test.index,)
-
-        return self
 
 
 # Backward compatibility
